pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gear-grinders"
version = "0.1.0"
description = "Agent orchestrator: backlog task -> pipeline -> PR"
readme = {text = "Agent orchestrator: backlog task -> pipeline -> PR", content-type = "text/plain"}
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "gitpython>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
gg = "gg.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
